# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Willamette Valley Map API
 * OpenAPI spec version: 0.1.0
 */
export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export type MarkerCategory =
  (typeof MarkerCategory)[keyof typeof MarkerCategory];

export const MarkerCategory = {
  winery: "winery",
  restaurant: "restaurant",
  farmstand: "farmstand",
} as const;

export interface Marker {
  id: number;
  name: string;
  note: string;
  category: MarkerCategory;
  lat: number;
  lng: number;
  website?: string | null;
  city?: string | null;
  createdAt: string;
}

export type CreateMarkerInputCategory =
  (typeof CreateMarkerInputCategory)[keyof typeof CreateMarkerInputCategory];

export const CreateMarkerInputCategory = {
  winery: "winery",
  restaurant: "restaurant",
  farmstand: "farmstand",
} as const;

export interface CreateMarkerInput {
  name: string;
  note: string;
  category: CreateMarkerInputCategory;
  lat: number;
  lng: number;
  website?: string | null;
  city?: string | null;
}

export type UpdateMarkerInputCategory =
  (typeof UpdateMarkerInputCategory)[keyof typeof UpdateMarkerInputCategory];

export const UpdateMarkerInputCategory = {
  winery: "winery",
  restaurant: "restaurant",
  farmstand: "farmstand",
} as const;

export interface UpdateMarkerInput {
  name?: string;
  note?: string;
  category?: UpdateMarkerInputCategory;
  website?: string | null;
  city?: string | null;
}

export interface MarkerStats {
  total: number;
  wineries: number;
  restaurants: number;
  farmstands: number;
}