# lib/api-zod/src/generated/types/marker.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Willamette Valley Map API
 * OpenAPI spec version: 0.1.0
 */
import type { MarkerCategory } from "./markerCategory";

export interface Marker {
  id: number;
  name: string;
  note: string;
  category: MarkerCategory;
  lat: number;
  lng: number;
  website?: string | null;
  city?: string | null;
  createdAt: string;
}