# lib/api-zod/src/generated/types/markerStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Willamette Valley Map API
 * OpenAPI spec version: 0.1.0
 */

export interface MarkerStats {
  total: number;
  wineries: number;
  restaurants: number;
  farmstands: number;
}