# lib/api-zod/src/generated/types/updateMarkerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Willamette Valley Map API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateMarkerInputCategory } from "./updateMarkerInputCategory";

export interface UpdateMarkerInput {
  name?: string;
  note?: string;
  category?: UpdateMarkerInputCategory;
  website?: string | null;
  city?: string | null;
}